+++ native/src/socle/heap/free_heap.c
@@ -1,6 +1,6 @@
 #include "heap.h"
 
-void free_heap(uint32_t cursor){
+void free_heap(uint64_t cursor){
     heap_element* element = (heap_element*)&heap[cursor];
     element->status = 2;
 }
