+++ native/src/socle/heap/write_heap.c
@@ -1,16 +1,16 @@
 #include "heap.h"
 
-uint32_t write_heap(void* data, uint32_t length)
+uint64_t write_heap(void* data, uint32_t length)
 {
     if ( length > BASE_RESERVATION_PER_THREAD ) {
         return RES_WRITE_MEMORY_ERROR;
     }
-    uint32_t available_memory = thread_reservation_limit - thread_reservation_cursor;
+    uint64_t available_memory = thread_reservation_limit - thread_reservation_cursor;
     if (available_memory < length) {
         reservation_heap();
     }
 
-    uint32_t cursor = thread_reservation_cursor;
+    uint64_t cursor = thread_reservation_cursor;
     heap_element* element = (heap_element*)&heap[cursor];
     element->status = 1;
     element->length = length;
